Refac: More Modularization with Better Dir Structure (#8)
