+++ .env.sample
@@ -1,4 +1,5 @@
 FLASK_ENV=development
+FLASK_APP=server
 DOMAIN=localhost:5000
 MOCK_CANVAS=false
 LOCAL_TIMEZONE=US/Pacific
