+++ .github/actions/bootstrap/action.yml
@@ -5,13 +5,17 @@ inputs:
     description: "Whether to install development dependencies"
     required: false
     default: "true"
+  python-version:
+    description: "Python version to use"
+    required: false
+    default: "3.10"
 runs:
   using: "composite"
   steps:
-    - name: Set up Python 3.10
+    - name: Set up Python ${{ inputs.python-version }}
       uses: actions/setup-python@v3
       with:
-        python-version: "3.10"
+        python-version: ${{ inputs.python-version }}
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
